pyproject.toml
[project]
name = "drone-log-bot"
version = "0.1.0"
description = "Telegram bot for tracking drone maintenance work"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.15,<4",
    "aiosqlite>=0.20,<1",
    "pydantic-settings>=2.7,<3",
    "structlog>=24.0,<25",
    "fpdf2>=2.8,<3",
]

[project.scripts]
drone-log-bot = "bot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
